Made mlmodel utils tests separate; extracted error string out to utils
